Adapted the code to the fact that input bits from the core station now, after rewiring the hardware, has an I2C port of its own. Port 0x26. Added port testing and logging of all I2C ports on startup.
Added logging on utility threads startup.
Fixed buglet in the code starting the 144 PA.
The code now turns off QRO path on click on the QSO button if the PA is not ready.

Changed font and padding on the log table.
Increased the morse transmitter thread's idle sleep time to 3 units.