Added Sun, Moon and local Aircraft tracking.
There is now a tooltip view connected to the Antenna control heading. Hovering that shows the current target stack

Sund and Moon tracking is now implemented. There is a Sun and Moon button for those two.
A new "Pop" button pops off the current top target.

Tracking of local aircraft can noe be done to help aircraft scatter experiments.
My locally installed FlightRadar24 receiver is being eavesdropped for data on
all received aircraft and is updated every 12 seconds within the server.
These data are then used to display aircraft symbols on the map showing their positions
and headings. Clicking an aircraft pushes a plane target which the antenna will follow until the signal is lost or a maximum of 30 minutes, whichever happens first.