Target stack popping fixes
Added some debug loggings.
Fixed many bugs where the initialization of target objects were
done wrongly, as the ttl parameter was sent to the update_in parameter,
This caused popping a target to suddenly removing the whole stack because
lots of targets had the ttl set to 90 seconds instead of one hour.
Changed the __init statements to use named parameters for better safety and readability.