Big refactoring: what is a mode?
The term "mode" has become ambigous. Historically it meant the mode of transmission, but as VHF/UHF has become more popular it has also been used for mode of wave propagation. This causes lots of confusion when reading the code.

This refactoring now  employs txmode for mode of transmission and propmode for mode of wave propagation.

The changes also goes into the column names in the database tales. Dis is not shown in any git diffs though.