Hovering the squares and stations problem fix.
When hovering squares or stations in the map the info on squares often got intermixed with the info on stations.
This change splits info to be shown upon hovering and info to show upon clicking.
The result is that worked stations in a displayed square is only shown if clicking.