Improved handling of augmented locators.
Augmented locators are looked up in the callbook table which collects data from PSKReporter.
When adding a QSO, and there is a locator given in the QSO,
the callbook table is looked up for locators that:
    matches the callsign
    matches the first 4 characters in the given locator
If more than one match is found, the latest one entered into the callbook table is used, and entered into the separate column in the log.

If there is no given locator, the callbook table is looked up for locators that matches just the callsign, and if there are more than one match,
the most specific locator is used.

Then, the augmented locator is used for displaying the location in the map.
The log also displays the augmented locator, but iy is not used for producing contest logs.