Edit logged QSO:s
Sometimes wrong data gets into the log. Now QSO:s may be edited. This commit contains the changes needed.

Clicking the edit symbol on the rightmost column brings up tha data for that row into
the second, editable row. When committed, the changed data will not add a new QSOm rather the QSO brought up will be updated in the database and in the log table on screen.