Add validation function for ADIF data fields
A new function is_valid() was added to the ADIF implementation for validating the data in a field with respect to the ADIF specification. This function provides comprehensive data validation for different data types according to the ADIF specifications. As part of this change, existing validation code in parse_adi() and write() functions were replaced with calls to this new function.