Add distance attribute to Target class and auto-tracking limitation
A 'distance' attribute has been added to the Target class initialization in target_tracking.py, and its value is now passed when creating a StationTarget. Also, an 'auto' parameter has been introduced to the az_track_station method in azel.py to prohibit auto-tracking if the distance to the target is less than 100 km. This parameter is used in clientmgr.py where the az_track_station method is called.