Refactor 'add_mh_on_map' method in clientmgr.py
The 'add_mh_on_map' method is updated to include more locators. The previous implementation only added single locator at a time. This clashes heavily with the changed implementation in the javascript code which replaces rects on view with the incoming in order to be able to change the locator precision without reloading.
Now it fetches all current locators, appends the new one, and then pushes these locators to the map.