Bug fix: Update default argument handling in get_contest_times
The original implementation used a mutable default argument (today=datetime.now()) that introduced a bugs causing the contest date calculation to return the latest contest prior to the date and time the stnmate server was started . This update changes the argument to None, allowing current date to be assigned within the function if no explicit value is provided, and thus also make the contest date returned be the kast contest prior to the invocation of the function, not the start of the server, which tends to be run for weeks and months. A long standing and elusive bug that never showed up when trying to debug the code.