Bug fix: Make new antenna targets be effective immediately when inroduced. This bug was caused by the antenna control loop sleeping a calculated time while the antenna was rotatinf. This sleep has now been made abortale.
A new AbortableSleep class was introduced, allowing the sleep operation during azimuth rotation to be interrupted if needed. The azimuth rotation logic was adjusted to utilize this new feature and the logic for azimuth rotation has been clarified with comments.

The 'CCW_BEARING_STOP' degree value was also updated due to a change of the mechanical calibration azimuth.